% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_engine.R
\name{call_model}
\alias{call_model}
\title{call_model}
\usage{
call_model(model_name, engine = "parametric")
}
\arguments{
\item{model_name}{name of model}

\item{engine}{select either parametric or non parametric model}
}
\description{
call model available in inst/models
all model should have been generated manually or with the help of this package.
Supported extension are C++ and Rcpp file.
}
