% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_input.R
\name{create_model}
\alias{create_model}
\title{create_model}
\usage{
create_model(
  type,
  constant = c("micro", "macro"),
  cov = NULL,
  engine = "nonmem"
)
}
\arguments{
\item{type}{prespescified type of model.}

\item{constant}{specify if the model should be parametrized using micro of macro constant}

\item{cov}{is a list containing parameters and covariates relationship and beta value}

\item{engine}{nonmem by default. as of version 1.0.0 only nonmem engine is supported}
}
\description{
create model is the generic function that allow user to very easily create a new model file
}
