% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_translate.R
\name{pk_data_translate}
\alias{pk_data_translate}
\title{pk_data_translate}
\usage{
pk_data_translate(
  id = NULL,
  time,
  evid = NA,
  cmt = 1,
  amt,
  rate = NULL,
  dv,
  mdv,
  cov = NULL,
  route = "PO"
)
}
\arguments{
\item{id}{vector of patient identity}

\item{time}{vector containing all time of drug intake and observation}

\item{evid}{event id (1 = administration, 0 = drug sample) non mandatory parameters, will be calculated based on amt and dv}

\item{cmt}{number of compartment of observation. cmt is set to 1 by default (central compartment)}

\item{amt}{amount of drug given}

\item{rate}{infusion rate, set to 0 if route is Per Os (PO)}

\item{dv}{dependent variable (concentration observed)}

\item{mdv}{missing dependent variable (0 observation record DV is not missing, 1 = dose record or other time varying record, but dv is missing)}

\item{cov}{list of covariate and their values}

\item{route}{correspond to the route of administration, is set to PO by default}
}
\description{
These function are here to translate dataset as an in put to mapbayr
}
